+++ packages/legacy/core/App/screens/Chat.tsx
@@ -14,11 +14,11 @@ import { Text } from 'react-native'
 import { GiftedChat, IMessage } from 'react-native-gifted-chat'
 
 import { isPresentationReceived } from '../../verifier'
+import InfoIcon from '../components/buttons/InfoIcon'
 import { renderComposer, renderInputToolbar, renderSend } from '../components/chat'
 import { renderActions } from '../components/chat/ChatActions'
 import { ChatEvent } from '../components/chat/ChatEvent'
 import { ChatMessage, ExtendedChatMessage, Role } from '../components/chat/ChatMessage'
-import InfoIcon from '../components/misc/InfoIcon'
 import { useNetwork } from '../contexts/network'
 import { useStore } from '../contexts/store'
 import { useTheme } from '../contexts/theme'
